+++ src/TwigComponent/tests/Unit/ComponentTemplateFinderTest.php
@@ -53,6 +53,8 @@ final class ComponentTemplateFinderTest extends TestCase
             'foo/bar.html.twig',
             'bar/foo/bar.html.twig',
             'foo/foo/bar.html.twig',
+            'foo/qux/index.html.twig',
+            'foo/foo/baz/index.html.twig',
         ];
         $loader = $this->createLoader($templates);
         $finder = new ComponentTemplateFinder($loader, 'foo');
@@ -60,6 +62,9 @@ final class ComponentTemplateFinderTest extends TestCase
         $this->assertEquals('foo/bar.html.twig', $finder->findAnonymousComponentTemplate('bar'));
         $this->assertEquals('foo/foo/bar.html.twig', $finder->findAnonymousComponentTemplate('foo:bar'));
         $this->assertEquals('foo/foo/bar.html.twig', $finder->findAnonymousComponentTemplate('foo:bar'));
+
+        $this->assertEquals('foo/qux/index.html.twig', $finder->findAnonymousComponentTemplate('qux'));
+        $this->assertEquals('foo/foo/baz/index.html.twig', $finder->findAnonymousComponentTemplate('foo:baz'));
     }
 
     private function createLoader(array $templates): LoaderInterface