+++ src/TwigComponent/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 -  Add option `profiler.collect_components` to control component data collection
    in the profiler (enabled in debug mode by default)
+-  Add support for using directory name as component name for anonymous components
 
 ## 2.30